(#207) Move builtin.py to 'cmd' folder
