+++ src/cmd/builtin.py
@@ -9,16 +9,16 @@ import urllib.request
 import discord
 import requests
 
-from . import const
-from . import emoji
-from .config import BackgroundEvent
-from .config import Command
-from .config import Reaction
-from .config import bc
-from .config import log
-from .quote import Quote
-from .reminder import Reminder
-from .utils import Util
+from .. import const
+from .. import emoji
+from ..config import BackgroundEvent
+from ..config import Command
+from ..config import Reaction
+from ..config import bc
+from ..config import log
+from ..quote import Quote
+from ..reminder import Reminder
+from ..utils import Util
 
 
 class BuiltinCommands:
